+++ src/apps/main/network/get-file-mirrors.ts
@@ -1,4 +1,4 @@
-import { Mirror } from './types';
+import { Mirror } from './requests';
 
 type Props = {
   bucketId: string;
